The hardware has 1MB of ram with working DAT to page it in. The pages are changed from 4kB to 8kB to be more like what NitrOS9 level 2 wants.
There is a working task register, and the total possbile RAM
is 2MB, with a 5-bit task register to extend the addess bus to
21 bits.

USB proof-of-concept code works; there is one serial port
and one vendor channel for buld data, but they don't do
anything useful yet.

There are lots of code tweeks, mostly random bits that I
found/fixed along the way.