A big load of further C++-ifying. The stronger checking looks like it cleaned out a few glitches. The code has been split into a couple more files as a part of this. The work is incomplete.
The console "fake" serial seems to work, but I get occasional doubled
characters on input when in IRQ mode. I have no idea if the bug is
on the 6809 side or the pico side.

The trace facility is a casualty of the C++-ification. It's a mess
that I'd like to rewrite.

There are some programs in the chunk/ directory that are building
towards a library and test suite.